+++ src/validators/some.ts
@@ -63,6 +63,10 @@ export function some(...validators) {
             return cur;
           }
 
+          if (Array.isArray(acc) && Array.isArray(cur)) {
+            return [...acc, ...cur];
+          }
+
           return deepMerge(acc, cur);
         },
         undefined,